chore(refactor): Better package structure
